RTE Solver Major Updata
Introduced the fully implicit (FI) scheme and the fourth-order
semi-implicit (SI) scheme that have much better accuracy and are times
more faster than the FE scheme. Tested them on various cases.

Fixed a bug with the export dialog not hiding the progress bar properly.
Had to change the way things worked with export.