Improving logic of baseline classes and of the parameter-setting
Create a separate abstract class for horizontally-adjustable baselines.
The FlatBaseline and LinearBaseline now both inherit from this class.
The reason to do this was a logical error when the LinearBaseline
extends a FlatBaseline and therefore it is a FlatBaseline (although 
it isn't).

Included a fitToNegative method in the Baseline class accepting 
AbstractData arguments. This is specifically needed to calculate the 
baseline of e.g. pulse diode data, which can't be treated neither as 
a HeatingCurve nor as an ExperimentalData.

Made an override of the listedKeywords() method instead of 
listedTypes()

Removed the constant values from the optimisationVector() method and
replaced explicitly setting bounds in this method by a call to the 
overriden set(int, double, NumericPropertyKeyword) method from 
ParameterVector, which takes the default bounds for the keyword from
the default loaded set of parameters.

Intoduced a utility boundsFrom method to the Segment.class to extract
the bounds from a default instance of the NumericProperty corresponding
to a given NumericPropertyKeyword