Improvements to handling of experimental datasets
ExperimentalData now stores a value of the half-time, which is used 
for accurate calculation of the Range bounds. The half-time is updated
using a DataLsistener addead to each new ExperimentalData object.

The Range objects will now perform checks when asked to set a new 
minimum or maximum, by verifying that the new value is contained within
a dynamically calculated range interval, which is based on the half-time
value of the main experimental dataset.

Removed pop-up windows appearing (uncontrollably) when the half-time
calculation failed, the output stream is now re-directed to System.err.

Replaced cubic normal splines in InterpolatorDataset with Akima splines.
This is not ideal, but at least will yield better results to datasets 
with rapidly varying second derivatives. In future, this needs to be 
replaced by a procedure producing uniformly spaced data to which LOESS
interpolation is applied.

Changes listedTypes in Metadata.java to listedKeyword(), which reduces 
the complexity of the code.