Improvements in exporting
Fixed a bug in the Exporter.java when the files were saved in a new 
format but with the default extension (e.g. html as csv). This was fixed
by manually replacing the extension substring with the selected export 
extension.

Modified the ResultTableExporter to correctly print the properties in 
a .csv table (both value and error, which are delimited by +/-) using
the String.format() functionality. 
Unfortunately, it was not possible to align the values in the columns
directly below the header string, but, perhaps, in a future release 
that could be fixed.

Fixed the directoryQuery() method in ExportDialog, which previously
incorrectly selected a file in the current directory, and not that 
directory itself. This had led to duplication of a directory (e.g.
if 'Documents' was selected, this would created a 'Documents' folder 
inside, and the export path would be 'Document/Documents/.../'.