Improvements in XML handling
The xml format was changed in such a way that xml elements declaring 
a 'default-search-variable' attribute will be marked as 'optimisable',
even if the value of that variable is 'false'. Therefore, there are now
two distinct states: isDefaultSearchVariable() and isOptimisable(). 
Any element that should be allowed to be optimisable should have a 
'default-search-variable' attribute. Those elements that declare 
this property as 'true' will have their search flag enabled by default.