Fixed property transforms and problem statements
Introduced a new StickTransform class, which prevents properties from
taking values outside a given region by returning their values back to 
the boundary (so that the property 'sticks' to the boundary). This has
proven to yield better (in terms of possible optimiser bias) results
compared to the previously used ATANH transform.

Changed the ATANH transform to ABS in NonlinearProblem.

Converted listedTypes() to listed Keywords()

Changed default baseline for all problem statements to LinearBaseline 
(previously - FlatBaseline). Fixed wrong baseline setting in the copy
constructor when the listeners were not notified of this, which caused
a loss of search parameters (BASELINE_INTERCEPT and BASELINE_SLOPE could
have gone missing).

Removed calls to calculated half-time in ThermalProperties, replacing 
them with a getter instead.