Re-arranged handling of Flags and NumericProperties
Introduced the optimisable boolean in NumericProperty. This boolean
indicates whether this property can be optimised.

Flag.allFlags() now returns a list of only those flags which 
correspond to numeric properties with isOptimisable() = true

Removed Flag.allProblemIndependentFlags() due to being obsolete

Intoduced dimensionDelta in NumericProperty to account for a possible
change of format of the test temperature (for example, conversion between
K to deg. C). This dimensionDelta is directly added to the value returned
by valueInCurrentUnts.

Fixed errorInCurrentUnits throwing typecast exceptions due to incorrect
casting to double. Now this works with any type of Numbers including ints.

Removed valueOutput() and errorOutput() (formatting is now a prerogative of 
a separate NumericPropertyFormatter)

Removed formatting methods from NumericProperties.

Formatting is now handled by NumericPropertyFormatter, which is a type
of AbstractFormatter that accepts NumericProperty objects. There are three
formats available to format any property: a standard decimal format with
fixed number of digits after the decimal point, an integer format for 
properties with integer values, and a newly introduced ScientificFormat,
which renders values using an html output, e.g.: <html> 10 &times; 5<sup>2</sup></html>.
This is then converted to a scientific notation. Works with JTables.