Optimiser update
Adds a MAX_FAILED_ATTEMPTS field, which represents acceptable number of 
retries when the optimiser refuses to accept the uphill step and tries
to re-configure itself to arrive at a downhill step. Previously this
was only possible for the LMOptimiser, however, in the new logic, even
the Hessian-based optimisers can now do this -- the re-configuration
is achieved by reseting the Hessian matrix to identity. In future releases,
this can probably be merged with LMOptimiser to reduce duplications.

The GradientBasedOptimiser now hosts a dx() method for calculating the 
gradient step, callable by all subclasses. Previously, the same code 
was inserted in different optimisers. In addition, it failed miserably
for properties which could reach 0.0, since, in that case, the gradient
step was calculated as a factor of the property and the GRADIENT_RESOLUTION
(either low- or high, depending on whether the property is discrete). This
caveat has been fixed by replacing the 0.0 value with the maximal property
value.

The IterativeState, a superclass for all paths, now stores the current
value of the cost function and the set of parameters corresponding to 
that cost function value. This is needed in order to be able to revert
back to a 'safe' set of parameters.

The maximum number of failed attempts for the LMOptimiser was reduced
to 4. At each successful attempt, the current state is stored in the 
path. 

listedTypes -> listedKeywords

All flags were removed from being listed in the PathOptimiser. The 
override for the data() method was removed.