Updated Calculation and SearchTask
Modified Calculation.setStatus() to return false in cases when a 
change of status is not appropriate (for example, when the Calculation
is being forced to run, although it is not ready). Contrary to the previous
implementation, the method does NOT return false when the status being
set is the same as the one already assigned. This functionality has been
transferred to SearchTask.setStatus() instead; only there this is needed
to avoid triggering the listeners without actually updating anything.

In SearchTask, a field has been added to store the global-best iterative
state. At the end of execution, it is checked whether the converged 
state is actually worse than the global-best -- if so, the global-best
set of parameters are taken instead, and the task is re-calculated.