Fixed result processing
Remove ResultOperation, as this was more confusing than it actually solved
any problems. The ResultProcessing has now changed to ResultStatistic.

The ResultStatistic has a method process(), which shuffles the input 
array of results and maps it to a Map<NumericPropertyKeyword, List<Double>>,
i.e. it collects all values of properties and group them by the property
keyword. This map is then used to calculate the averages and the errors 
of the average properties. The output is assigned strictly according to 
the specified ResultFormat, i.e. the order in which properties appear
are taken from the ResultFormat instance.

An error associated with the calculation of the standard deviation of the 
mean was corrected. This is associated with the fact that, such standard
deviation was calculated as bias-uncorrected stdev of the population, i.e.
it was larger than the stdev of the mean by sqrt(n-1). This has now been
corrected.

The error value now represents half of the span of the confidence interval
calculated at a confidence level of 0.95. This is achieved by calculating
the inverse cumulative distribution function of the t-distribution, for
the number of degrees of freedom equal to the input array size (number of 
results). The qunatile of the t-distribution is then multiplied by the 
stdev of the mean.

The AverageResult.calculate() method was modified to allow calculating 
the mean values and errors associated with averaging of integer-type 
results (the error is then rounded up to an integer value). Additionally,
the sqrt(...) operation was removed for the calculation of stdev from
the variance, as the ResultStatistics now has the full calculation of errors.