Charts, markers, and mouse events - update
Introduced interactive ValueMarkers in Charts, which mark the boundaries
of the search range. The markers are now highlighted whenever the 
cursor appears in the direct vicinity. They also respond to mouseDragged
events now, which not only changes the rendering of those items in the 
chart, but simultaneously changes the Range object of the currently
selected task's ExperimentalData. The markers can only be dragged within
pre-defined 'sanity' regions. For example, attempting to drag them at 
t > tmax will set the marker position to the top boundary (= tmax).
The markes are automatically updated whenever the Range's lower or upper
bounds change -- including during the optimisation process (in real time).

The changes to the ExperimentalData's range is reflected in the range
JFormattedTextFields appearing in the ChartToolbar. Now these elements
are specified in their own class, RangeTextFields. The values in the 
textfields are synchronised with both the ExperimentalData's Range and 
the ValueMarkers on the the Charts, so that each edit to the ValueMarker
and to the Range will be reflected in the text fields. However, if the
textfields are edited themselves, the Range will only be updated when
the user clicks to Set Range Button (formerly Limit Range To). In addition,
typing an illegal value in the textfields now results in that value 
being discarded. Only numbers within the intervals specified in the 
Range class are allowed. Finally, the text fields display the range 
in milliseconds with one digit after the decimal point (this is additionally
highlighted by the 'ms' suffix). 

When the Set Range button is pressed, the user is given three different 
options now: to update range to All tasks, just to the current one, and
to discard changes. The pop-up window now has text wrapping set by the
html tags. The property values displayed are formatted according to the 
same convention used for the range textfields.