Data loading and Problem/Scheme Changes - Concurrency
Loading numerical pulse information is now handled through a combination
of a Cached Thread Pool, where requests update properties based on the
loaded information are submitted, and a Single Thread Executor, 
which manages the read requests. This avoids deadlocks.

ProblemStatementFrame: Some re-shuffling of code to reduce complexity 
-> new methods changeProblems(), changeSchemes(). 

Three Cached Thread Pools now manage changing problems, schemes, and
updating properties -- this greatly improves the performance of processing 
many tasks (hundreds) and ensures that processing is done in the correct
order (that also avoids ConcurrentModificationException!). To ensure 
correct operation, each time when a heap of tasks is required to be processed,
the pools use the invokeAll() method, preserving the operation lock until
processing is complete. The only exception here is the property-update
pool, which simply accepts submitted callables and doesn't use invokeAll().