Tables and renderers - changes
merge() and group() have been removed from the ResultTable, as they 
really should be included in the table model instead. 

The new merge() and group() methods introduced in the ResultTableModel
now handle the results directly, instead of invoking getValueAt(). The
merge logic proceeds as follows: first, the result list is sorted in 
order of ascending ids. Then, on each element of that list, minus the
skip list, the method group is invoked, which creates a group of results
matching the following criteria: (a) these results test temperature should
be within a specified tolerance (specified by the user) and (b) the 
minimum difference between their id value must be less-or-equal-to unity.
This ensure that only a sequence of shots is averaged; that avoids 
averaging over the heating and cooling phase, for example. After generating
a result from that group, which can either be an individual result, if 
the group size is one, or an average results, these results forming the 
group are then added to the skip list, and thus excluded from being processed
later on. 

TableRenderer -> changed getTableCellRendererComponent() to always have 
task ids and status centred in the respective cells. 

AccessibleTableRenderer -- removed colouring of Properties and Property
Holders. 

Changed the default AbstractFormatter of the JFormattedTextFields generated
by the NumericPropertyRenderer to NumericPropertyFormatter. If the property
is not editable, a JLabel is created with its text set to the getText() of 
the JFormattedTextField (thus, the formatting is preserved).