Toolbars and dialogs
Plotting requests in ProblemToolbar are now handled via an ExecutorService to avoid
concurrent modification.

FormattedInputDialog now formats the properties according to the 
NumericPropertyFormatter, thus maintaining consistency. This means that 
checks for isEditValid() are no longer necessary, since invalid edits
will be rejected by the formatter anyway. Finally, the method value()
had to be adjusted, as with the new formatter, the return type of the
ftf.getValue() changed to NumericProperty.