Preview plotting
Previously, a XYSplineRenderer was used to interpolated between points
for the result plot. However, it did not yield good results, and in most
tricky cases it generated a bizarre curve which did not do the interpolation
job well. One of the reasons for this could be a poor implementation of
natural cubic splines in JFreeChart -- the other could be the general 
problem associated with badly fitting data with rapidly varying second
derivatives. Regardless of that, the new version now constructs a spline
interpolation using ApacheCommonsMath, and, for a fixed sample size of 
100 points, calculates the discrete points of the curve, which it then
renders using a XYLineAndShapeRenderer. In the vast majority of cases,
this yields better results. It is worth mentioning that, the routine
tries to use the LoessInterpolator first, which will give generally 
reasonable results for a dataset with noise (or random error). If this fails
(if an exception is thrown), then it switches to AkimaSplineInterpolator.
The routine will exit if the points are not strictly increasing. In future,
this needs to be changed to generate a set of points with uniform spacing,
which would then be interpolated.