Minor Fixes
- Fixed pulse visualisation. Now it takes parameters directly from 
the PulseShape, which is initialised by the DiscretePulse. Moreover,
it only plots points at discrete time steps defined by the Grid.
Therefore, the user should have an idea as to how detailed the pulse 
correction is.
- Adjusted the pulse resolution allowed for 2D models. This had to be
done since 2D calculations are more demanding in terms of the number of
pulse points. Previously these resulted in incorrect Tinf values.